requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
opencv-python